# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Neuralis AI — Multi-feature AI platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  mode: string;
  createdAt: string;
  updatedAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  mode: string;
  createdAt: string;
  updatedAt: string;
  messages: OpenaiMessage[];
}

export interface CreateOpenaiConversationBody {
  title: string;
  mode?: string;
}

export interface UpdateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface SendOpenaiVoiceMessageBody {
  /** Base64-encoded audio data */
  audio: string;
}

export type GenerateOpenaiImageBodySize =
  (typeof GenerateOpenaiImageBodySize)[keyof typeof GenerateOpenaiImageBodySize];

export const GenerateOpenaiImageBodySize = {
  "1024x1024": "1024x1024",
  "1536x1024": "1536x1024",
  "1024x1536": "1024x1536",
} as const;

export interface GenerateOpenaiImageBody {
  prompt: string;
  size?: GenerateOpenaiImageBodySize;
  style?: string;
}

export interface GenerateOpenaiImageResponse {
  b64_json: string;
  prompt: string;
}

export interface OpenaiStats {
  totalConversations: number;
  totalMessages: number;
  imagesGenerated: number;
  studySessions: number;
  healthConsultations: number;
}

export interface OpenaiError {
  error: string;
}

export interface WeatherCurrent {
  temp: number;
  feelsLike: number;
  humidity: number;
  windSpeed: number;
  condition: string;
  description: string;
  icon: string;
  visibility: number;
  pressure: number;
  uvIndex: number;
}

export interface WeatherForecastDay {
  date: string;
  dayName: string;
  tempMin: number;
  tempMax: number;
  condition: string;
  icon: string;
  precipitationChance: number;
}

export interface WeatherResponse {
  city: string;
  country: string;
  current: WeatherCurrent;
  forecast: WeatherForecastDay[];
  aiInsight: string;
}

export interface WikipediaSearchResult {
  title: string;
  excerpt: string;
  url: string;
  /** @nullable */
  thumbnail?: string | null;
}

export interface WikipediaSummary {
  title: string;
  extract: string;
  aiSummary: string;
  url: string;
  /** @nullable */
  thumbnail?: string | null;
  relatedTopics: string[];
}

export interface HealthcareConversation {
  id: number;
  title: string;
  specialty: string;
  createdAt: string;
  updatedAt: string;
}

export interface HealthcareConversationWithMessages {
  id: number;
  title: string;
  specialty: string;
  createdAt: string;
  updatedAt: string;
  messages: OpenaiMessage[];
}

export interface CreateHealthcareConversationBody {
  title: string;
  specialty?: string;
}

export interface SendHealthcareMessageBody {
  content: string;
}

export interface AnalyzeSymptomsBody {
  symptoms: string[];
  /** @nullable */
  age?: number | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  duration?: string | null;
}

export interface SymptomCondition {
  name: string;
  likelihood: string;
  description: string;
}

export interface AnalyzeSymptomsResponse {
  disclaimer: string;
  possibleConditions: SymptomCondition[];
  recommendations: string[];
  urgencyLevel: string;
}

export interface StudySession {
  id: number;
  title: string;
  subject: string;
  flashcardCount: number;
  quizCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface Flashcard {
  id: number;
  sessionId: number;
  front: string;
  back: string;
  difficulty: string;
  createdAt: string;
}

export interface QuizQuestion {
  id: number;
  sessionId: number;
  question: string;
  options: string[];
  correctIndex: number;
  explanation: string;
  createdAt: string;
}

export interface StudySessionWithContent {
  id: number;
  title: string;
  subject: string;
  flashcardCount: number;
  quizCount: number;
  createdAt: string;
  updatedAt: string;
  flashcards: Flashcard[];
  quizQuestions: QuizQuestion[];
}

export interface CreateStudySessionBody {
  title: string;
  subject: string;
}

export interface GenerateFlashcardsBody {
  text: string;
  count?: number;
}

export interface GenerateQuizBody {
  text: string;
  count?: number;
  difficulty?: string;
}

export type SummarizeTextBodyStyle =
  (typeof SummarizeTextBodyStyle)[keyof typeof SummarizeTextBodyStyle];

export const SummarizeTextBodyStyle = {
  brief: "brief",
  detailed: "detailed",
  bullet: "bullet",
  academic: "academic",
} as const;

export interface SummarizeTextBody {
  text: string;
  style?: SummarizeTextBodyStyle;
}

export interface SummarizeTextResponse {
  summary: string;
  keyPoints: string[];
  wordCount: number;
}

export type ExplainConceptBodyLevel =
  (typeof ExplainConceptBodyLevel)[keyof typeof ExplainConceptBodyLevel];

export const ExplainConceptBodyLevel = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
  expert: "expert",
} as const;

export interface ExplainConceptBody {
  concept: string;
  level: ExplainConceptBodyLevel;
  /** @nullable */
  subject?: string | null;
}

export interface SubjectStat {
  subject: string;
  sessionCount: number;
}

export interface StudyStats {
  totalSessions: number;
  totalFlashcards: number;
  totalQuizzes: number;
  subjectBreakdown: SubjectStat[];
}

export type GetWeatherParams = {
  /**
   * City name to get weather for
   */
  city: string;
  /**
   * Temperature units (metric=Celsius, imperial=Fahrenheit)
   */
  units?: GetWeatherUnits;
};

export type GetWeatherUnits =
  (typeof GetWeatherUnits)[keyof typeof GetWeatherUnits];

export const GetWeatherUnits = {
  metric: "metric",
  imperial: "imperial",
} as const;

export type SearchWikipediaParams = {
  query: string;
  limit?: number;
};

export type GetWikipediaSummaryParams = {
  /**
   * Wikipedia article title
   */
  title: string;
};